pyproject.toml
[project]
name = "wechat-mcp"
version = "0.1.0"
description = "微信公众号MCP服务器"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "diskcache>=5.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setup.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']
